GDTerminal/requirements.txt
dominate==2.6.0
future==0.18.2
imageio==2.9.0
imutils==0.5.4
numpy==1.20.1
opencv-python==4.5.1.48
pandas==1.2.3
Pillow==8.1.2
scikit-image==0.18.1
scipy==1.6.2
torch==1.8.1
torchfile==0.1.0
torchvision==0.9.1
visdom==0.1.8.9
